% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bardecph.R
\name{bardecph}
\alias{bardecph}
\title{Barplot decroissant en \%
Trace un barplot d'une variable factorielle, axe des y en \%, classement des items en décroissant}
\usage{
bardecph(varx, titre)
}
\arguments{
\item{varx}{variable a traiter (factorielle)}

\item{titre}{Titre du graphique}
}
\value{
un graphique
}
\description{
Barplot decroissant en \%
Trace un barplot d'une variable factorielle, axe des y en \%, classement des items en décroissant
}
\examples{
bardecph(iris$Species, "Espece")
}
