% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barouiph.R
\name{barouiph}
\alias{barouiph}
\title{Barplot avec barres d'erreur d'une valeur}
\usage{
barouiph(
  dfx,
  varx,
  testx,
  valx = "oui",
  titre = "",
  stitre = "",
  titx = "",
  tity = "\%",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{dfx}{data.frame}

\item{varx}{variable à représenter (factoriel) y}

\item{testx}{Variable de tri (factoriel) x}

\item{valx}{valeur de test à présenter}

\item{titre}{Titre du graphique}

\item{stitre}{Sous-titre du graphique}

\item{titx}{titre (optionnel) de l'axe x}

\item{tity}{titre (\% par défaut) de l'axe y}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
graphique
}
\description{
Trace un barplot avec les intervalles de confiance pour une valeur donnée
d'une variable factorielle selon les classes d'une variable factorielle de
classification.
}
\examples{
data(patients)
barouiph(dfx = patients, varx =escarre, testx =sexea, valx = "oui", titre = "Sexe",angle = 0)

}
