% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barpercentph.R
\name{barpcph}
\alias{barpcph}
\title{Barplot en pourcentage}
\usage{
barpcph(dfx, varx, vary, tit = "", stit = "", cap = "", tx = "", angle = 0)
}
\arguments{
\item{dfx}{Tibble}

\item{varx}{Variable de tri, (axe X)}

\item{vary}{Variable Y (\%)}

\item{tit}{Titre}

\item{stit}{Sous titre}

\item{cap}{Texte de légende}

\item{tx}{Titr de l'axe X}

\item{angle}{Angle des étiquettes de l'axe X}
}
\value{
graphique
}
\description{
Affiche un bar-plot de deux variables discrètes, avec la variable en Y présentée en \% pour chaque niveau de la variable X.
}
\examples{
barpcph(dfx = patients, varx = sexea, vary = admission, tit="Admission selon le sexe", stit ="Démonstration", cap="Démonstration sur Admission selon le sexe", tx="Sexe", angle = 0)


}
