% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barsimpleph.R
\name{barsimpleph}
\alias{barsimpleph}
\title{Barplot simple en \%
Trace un barplot d'une variable factorielle, axe des y en \%}
\usage{
barsimpleph(dfx, varx, titre = "", stitre = "", capt = "", lab = "", angle = 0)
}
\arguments{
\item{dfx}{data.frame}

\item{varx}{nom de la variable à traiter (factorielle)}

\item{titre}{Titre du graphique}

\item{stitre}{Sous-titre du graphique}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle d'affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
un graphique
}
\description{
Barplot simple en \%
Trace un barplot d'une variable factorielle, axe des y en \%
}
\examples{
barsimpleph(dfx = patients, varx= admission,
titre = "Mode d'admission", stitre ="ICU",
capt ="Admission", lab = "aa", angle = 10)

}
