% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bashaut.R
\name{bashaut}
\alias{bashaut}
\title{bashaut}
\usage{
bashaut(xx)
}
\arguments{
\item{xx}{Une variable numérique}
}
\value{
deux nombres : la borne inf et la borne sup de l'IC
}
\description{
Calcul de l'intervalle de confiance à 95 \% de la moyenne d'une variable numérique (utilisé par barconfph)
}
\examples{
data ("patients")
bashaut(xx =patients$age)


}
