% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaup.R
\name{beaup}
\alias{beaup}
\title{Petit p bien presenté}
\usage{
beaup(varp, affp = 0)
}
\arguments{
\item{varp}{résultat du p}

\item{affp}{si TRUE renvoie "p = 0,005" sinon "0,005"}
}
\value{
beaup , chaine de caractère
}
\description{
Petit p bien presenté
}
\examples{
pp <- cor.test(iris$Sepal.Length, iris$Sepal.Width)
beaup(pp$p.value)


}
