% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicph.R
\name{bicph}
\alias{bicph}
\title{Affiche un nombre avec son intervalle de confiance}
\usage{
bicph(nn, bb, hh, pc = 3)
}
\arguments{
\item{nn}{nb brut}

\item{bb}{Borne basse de l'IC}

\item{hh}{Borne haute de l'IC}

\item{pc}{Nombre de chiffres affiche}
}
\value{
chaine de caractère
}
\description{
Affiche un nombre avec son intervalle de confiance
}
\examples{
bicph(nn = 55, bb = 22, hh = 77, pc = 3)
}
