% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiph.R
\name{bmiph}
\alias{bmiph}
\title{Coupe un vecteur BMI ou IMC (numeric) en facteurs avec les limites usuelles, en français ou en anglais (les limites sont différentes)}
\usage{
bmiph(bmi, lang = "fr")
}
\arguments{
\item{bmi}{BMI numérique}

\item{lang}{Langue de sortie (french or english (fr/eng)). default : french}
}
\value{
vecteur factoriel
}
\description{
Coupe un vecteur BMI ou IMC (numeric) en facteurs avec les limites usuelles, en français ou en anglais (les limites sont différentes)
}
\examples{
imc <- c(18,25,40,32)
bmiph(imc, "fr")
}
