% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnom.R
\docType{data}
\name{bnom}
\alias{bnom}
\title{Données (fictives & partielles) sur les escarres en réanimation. Noms corrects des variables.}
\format{
A data frame with 8 rows and 2 variables:
\describe{
\item{code}{Nom de la variable}
\item{nom}{Beau nom pour l'affichage}
}
}
\source{
\url{https://www.australiancriticalcare.com/article/S1036-7314(20)30349-0/fulltext}
}
\usage{
bnom
}
\description{
Données (fictives & partielles) sur les escarres en réanimation. Noms corrects des variables.
}
\keyword{datasets}
