% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debph.R
\name{debph}
\alias{debph}
\title{DEBPH}
\usage{
debph(param = FALSE)
}
\arguments{
\item{param}{TRUE si variables normales (moyenne et écart-type), FALSE si variables non normales (médiane et quartiles)}
}
\value{
affs (affichage des valeurs numériques)
}
\description{
Réglages de base pour les tableaux. À éxécuter au début du document Quarto. Chargement de quleques libraries.
}
\examples{
debph(param = FALSE)

}
