% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debutph.R
\name{debutph}
\alias{debutph}
\title{Début du travail}
\usage{
debutph(fich, nom = "xx")
}
\arguments{
\item{fich}{Fichier csv contenant les labels corrects des variables (colonne 'nom')}
}
\value{
data.frame
}
\description{
Lit un csv contenant les données, simplifie & normalise les titres des variables.
le vecteur \code{nom} contenant les beaux labels des variables pour les tableaux, graphiques etc. sera incorporé dans le tableau principal.
}
\details{
Le csv doit avoir les réglages par défaut (UTF-8, séparateur de champs : ",") soit les réglages par défaut de LibreOffice
.
Toutes les variables "caractère" sont transformées en facteur.
}
