% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figp.R
\name{figp}
\alias{figp}
\title{Dessin des crochets avec texte pour les p-value dans les graphiques}
\usage{
figp(pp, x1, x2, yy, pval, od = FALSE, h = 1.1)
}
\arguments{
\item{pp}{Un objet ggplot2}

\item{x1}{Numéro d'ordre de la première variable}

\item{x2}{Numéro d'ordre de la deuxième variable}

\item{yy}{Niveau vertical de la barre horizontale}

\item{pval}{p-value à afficher}

\item{od}{Décalage horizontal des crochets (si plusieurs niveaux)}

\item{h}{Écart barre horizontale-texte}
}
\value{
graphique
}
\description{
Dessin des crochets avec texte pour les p-value dans les graphiques
}
\examples{
fp <- patients |>
  drop_na(typemaladie, igs2) |>
  ggplot(aes(x = typemaladie, y = igs2)) +
  geom_boxplot()
figp(fp, x1 = 1, x2 = 2, yy = 95, pval = 0.01, od = FALSE, h = 2)

}
