% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmph.R
\name{glmph}
\alias{glmph}
\title{Tableau de regresion logistique}
\usage{
glmph(
  dfx,
  vart,
  vars,
  bnom,
  titre = "Analyse multivariée",
  lab = "",
  export = FALSE
)
}
\arguments{
\item{dfx}{nom du data-frame}

\item{vart}{nom de la variable dependante}

\item{vars}{vecteur des variables explicatives (garder l'ordre du data-frame)}

\item{bnom}{vecteur contenant les intitules corrects des variables (l'extraction des valeurs utiles est automatique)}

\item{titre}{titre du tableau}

\item{lab}{label}

\item{export}{si TRUE : exporte la table en csv?}
}
\value{
tableau
}
\description{
Trace le tableau d'une regression logistique avec les intitules vrais des variables.
}
\examples{
irisz <- iris
          irisz$zz <- as.factor(rep(c("a","b"),75))
          glmph(dfx = "irisz",vart = "zz",vars = c("Sepal.Width","Species"),bnom = names(irisz))

}
