% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histmultiph.R
\name{histmultiph}
\alias{histmultiph}
\title{Graphique en plusieurs histogrammes superposés, pratique pour bien visualiser les variations d'une distribution selon une modalité.}
\usage{
histmultiph(dfx, varx, varn, tit = "", stit = 0, titx = "", bin = 1)
}
\arguments{
\item{dfx}{data-frame ou tibble}

\item{varx}{Variable factorielle}

\item{varn}{variable numérique}

\item{tit}{Titre du graphique}

\item{stit}{valeur de p}

\item{titx}{Titre de l'axe horizontal}

\item{bin}{largeur d'une barre}
}
\value{
un graphique
}
\description{
Graphique en plusieurs histogrammes superposés, pratique pour bien visualiser les variations d'une distribution selon une modalité.
}
\examples{
histmultiph(dfx = iris, varx= Species,
varn = Sepal.Width, tit = "IRIS", stit = 0.002, titx = "Pétales (mm)", bin = 1)

}
