% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligfact.R
\name{ligfact}
\alias{ligfact}
\title{Ligne facteurs tableau descriptif}
\usage{
ligfact(x, nn)
}
\arguments{
\item{x}{vecteur de donnees}

\item{nn}{nom du vecteur}
}
\value{

}
\description{
Ligne facteurs tableau descriptif
}
\examples{
ligfact (iris$Species, "Especes")

}
