% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignum.R
\name{lignum}
\alias{lignum}
\title{Ligne numerique tableau descriptif}
\usage{
lignum(x, nn)
}
\arguments{
\item{x}{vecteur de donnees}

\item{nn}{nom du vecteur}
}
\value{

}
\description{
Ligne numerique tableau descriptif
}
\examples{
lignum (iris$Sepal.Length, "largeur petales")

}
