% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignumc.R
\name{lignumc}
\alias{lignumc}
\title{ligne numerique tableau comparatif}
\usage{
lignumc(nom, varp, trip)
}
\arguments{
\item{nom}{nom de la ligne}

\item{varp}{variable explicative (numerique)}

\item{trip}{variable expliquee (facteur)}
}
\value{
ligne complète avec p
}
\description{
ligne numerique tableau comparatif
}
\examples{
ligfc("petale",iris$Species,iris$Species)
}
