% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabrandom.R
\name{listrandph}
\alias{listrandph}
\title{Table de randomisation
Plusieurs centres, blocs de taille variable}
\usage{
listrandph(nbcent = 1, nbtrait = 2, nbcas)
}
\arguments{
\item{nbcent}{nombre de centres (default : 1)}

\item{nbtrait}{nombre de groupes (default : 2)}

\item{nbcas}{Nombre cas souhaites}
}
\value{
un csv global  + un csv par centre
}
\description{
Table de randomisation
Plusieurs centres, blocs de taille variable
}
\author{
Philippe MICHEL
}
