% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopph.R
\name{lollipph}
\alias{lollipph}
\title{lollipopph}
\usage{
lollipph(dfx, nom, tri = "xx", titre = "", capt = "x")
}
\arguments{
\item{dfx}{data-frame}

\item{nom}{colonne à exposer (factorielle)}

\item{tri}{niveaux de "nom" à mettre en évidence}

\item{titre}{titre}

\item{capt}{légende}
}
\value{
graphique
}
\description{
Tracé d'un graphique "lollipop" de distribution d'une variable factorielle avec possibilité de mise en évidence d'un ou plusieurs facteurs.
}
\examples{
lollipph(dfx = patients, nom = lieudevie1, 
tri = c("EHPAD","Maison de retraite"), 
titre = "Lieu de vie", capt = "Lieu de vie avant l'hospitalisation")

}
