% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moys.R
\name{meds}
\alias{meds}
\title{Présentation mediane  quantiles}
\usage{
meds(x)
}
\arguments{
\item{x}{variable numerique}
}
\value{
mediane (quartiles)
}
\description{
Présentation mediane  quantiles
}
\examples{
meds(patients$age)
}
