% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbequiph.R
\name{nb.equi.ph}
\alias{nb.equi.ph}
\title{Nb de sujet  Test d'équivalence proportions
Calcul du nb de sujet par groupe pour
un test d'équivalence sur des proportions.}
\usage{
nb.equi.ph(po, dl = 0.1, zalpha = 0.05, zbeta = 0.2)
}
\arguments{
\item{po}{proportion sur le groupe controle}

\item{dl}{deviation jugee comme acceptable (0.1)}

\item{zalpha}{risque alpha (0.05)}

\item{zbeta}{risque beta (0.2)}
}
\value{

}
\description{
Nb de sujet  Test d'équivalence proportions
Calcul du nb de sujet par groupe pour
un test d'équivalence sur des proportions.
}
\examples{
nb.equi.ph(po=0.5,dl = 0.1, zalpha = 0.05, zbeta = 0.2)

}
