% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbobsph.R
\name{nb.obs.ph}
\alias{nb.obs.ph}
\title{TNb de sujets pour un sondage
Calcul du nombre de cas necessaires pour une enquète obsevationelle
simple snas aucun test.
Si il y aplusiuers questions prendre px = 0.5}
\usage{
nb.obs.ph(px = 0.5, ex = 0.1, np = 1e+05)
}
\arguments{
\item{px}{Proportion estimé des réponses (0.5)}

\item{ex}{Marge d'erreur considere comme acceptable (0.1)}

\item{np}{Taile de la population totale (1e5)}
}
\value{

}
\description{
TNb de sujets pour un sondage
Calcul du nombre de cas necessaires pour une enquète obsevationelle
simple snas aucun test.
Si il y aplusiuers questions prendre px = 0.5
}
\examples{
nb.obs.ph(px = 0.5, ex = 0.1, np = 1e5)
}
