% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesp.R
\name{nesp}
\alias{nesp}
\title{Espaces superflus}
\usage{
nesp(x)
}
\arguments{
\item{x}{chaine de caractères}
}
\value{
chaine de caractères sans espaces superflus
}
\description{
Élimine les espaces en début et fin d'une chaîne de caractères.
}
\examples{
nesp("  Bonjour à   vous ")

}
