% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\docType{data}
\name{nn}
\alias{nn}
\title{Intitules corrects pour les donnees (fictives) contenues dans le fichier "patients" qui seront utilisees pour les tableaux par ex.}
\format{
A data frame with 8 rows and 1 variables:
\describe{
\item{nom}{Beaux noms des items}
}
}
\source{
\url{https://www.australiancriticalcare.com/article/S1036-7314(20)30349-0/fulltext}
}
\usage{
nn
}
\description{
Intitules corrects pour les donnees (fictives) contenues dans le fichier "patients" qui seront utilisees pour les tableaux par ex.
}
\keyword{datasets}
