% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{Donnees (fictives & partielles) sur les escarres en réanimation}
\format{
A data frame with 100 rows and 9 variables:
\describe{
\item{np}{numero ID}
\item{admission}{Mode d'admission}
\item{alite.7.j.av}{Alite plus de 7 jours avant (O/N)}
\item{age}{Age du patient}
\item{taille}{Taille}
\item{sexe}{Sexe}
\item{poids.ad}{Poids a l'admission}
\item{igs}{Score IGS2}
}
}
\source{
\url{https://www.australiancriticalcare.com/article/S1036-7314(20)30349-0/fulltext}
}
\usage{
patients
}
\description{
Donnees (fictives & partielles) sur les escarres en réanimation
}
\keyword{datasets}
