% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expk.R
\name{pexptabph}
\alias{pexptabph}
\title{Sortie d'un tableau gtsummary en tableau kableExtra (pour export pdf via LaTeX) sans adaptation à la page ou comme feuille d'un tableur ods. Longtable possible}
\usage{
pexptabph(dfk, exp = FALSE, nomfich = "export", nomsheet = "x", lg = FALSE)
}
\arguments{
\item{dfk}{objet gtsummary}

\item{exp}{booleen. FALSE : vers kableExtra, TRUE : vers ods}

\item{nomfich}{nom du classeur ods}

\item{nomsheet}{nom de la feuille dans le classeur ods}

\item{lg}{Booleeen. TRUE si sortie en longtable}
}
\value{
un tableau kbl prévu pour un export LaTeX
}
\description{
Sortie d'un tableau gtsummary en tableau kableExtra (pour export pdf via LaTeX) sans adaptation à la page ou comme feuille d'un tableur ods. Longtable possible
}
\examples{
zz <- gtsummary::tbl_summary(iris)
pexptabph(dfk = zz, exp = FALSE)

}
