% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyrph.R
\name{pyrph}
\alias{pyrph}
\title{Pyramide des ages
Trace une pyramide des ages simple.}
\usage{
pyrph(dfx, age, sexe, ff, mm, titre)
}
\arguments{
\item{dfx}{data.frame contenant les valeurs}

\item{age}{valeur des ages , idealement divise en classes}

\item{sexe}{variable factorielle binaire}

\item{ff}{valeur de la classe 'Feminin"}

\item{mm}{valeur de la classe "masculin"}

\item{titre}{Titre du graphique}
}
\description{
Pyramide des ages
Trace une pyramide des ages simple.
}
\examples{
aa <- runif(100,0,100)
          aa <- cut(aa,seq(0,100,10))
          sx <- c(rep("F",50),rep("M",50))
          df <- data.frame(aa,sx)
          pyrph(dfx = df, age = aa, sexe = sx, ff = "F", mm = "M", titre = "Pyramide des âges")

}
