% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloudph.R
\name{raincloudph}
\alias{raincloudph}
\title{raincloudph}
\usage{
raincloudph(df, vcat, vnum, titre = "", titcat = "", titnum = "", adj = 1)
}
\arguments{
\item{df}{dataframe}

\item{vcat}{variable de tri (catégorielle)}

\item{vnum}{variable à afficher (numérique)}

\item{titre}{titre du graphique}

\item{titcat}{titre de la variable catégorielle}

\item{titnum}{titre de la variable numérique}

\item{adj}{facteur d'ajustement de la courbe de densité (1 par défaut)}
}
\description{
raincloudph
}
\examples{
raincloudph(df = patients, vcat = admission, vnum = igs2, titre = "IGS2 vs admission", titcat = "Admission", titnum = "IGS 2", adj = 1)
}
