% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1ph.R
\name{tab1ph}
\alias{tab1ph}
\title{Tableau descriptif}
\usage{
tab1ph(dfx, titre = "Tableau descriptif", label = "tabd")
}
\arguments{
\item{dfx}{Data.frame to explore}

\item{titre}{title of the table}

\item{label}{label of the table for RMarkdown or LaTeX}
}
\value{

}
\description{
Tableau descriptif
}
\examples{
tab1ph(iris, "Table 1")
}
