% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_oo.R
\name{tab_oo}
\alias{tab_oo}
\title{Tableau Odds-Ratio, Risque Relatif & nb à traiter}
\usage{
tab_oo(df, intervention, evenement)
}
\arguments{
\item{df}{a tibble}

\item{intervention}{variable d'entree}

\item{evenement}{Variable de sortie (diagnostic)}
}
\value{
a table
}
\description{
Tableau Odds-Ratio, Risque Relatif & nb à traiter
}
\examples{
tab_oo(patients, "sexe", "escarre")
}
