% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabdeuxph.R
\name{tabdeuxph}
\alias{tabdeuxph}
\title{barplot deux variables factorielles
Barplot de deux variables factorielles representees en \% de la deuxieme dans chaque classe de la premiere}
\usage{
tabdeuxph(varx, vart, titre = "", stit = "", ang = 90)
}
\arguments{
\item{vart}{variable a decouper en \%}

\item{titre}{titre du graphique}

\item{stit}{soustitre du graphique}

\item{ang}{angle des legendes en x}
}
\value{
graphique
}
\description{
barplot deux variables factorielles
Barplot de deux variables factorielles representees en \% de la deuxieme dans chaque classe de la premiere
}
\examples{
aa <- c("a","a","b","c")
          bb <- c("1","2","3","3")
          tabdeuxph(aa,bb,"essai","test",ang=80)
}
