% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabph.R
\name{tabph}
\alias{tabph}
\title{Réglages pour un tableau généré par gtsummary (avec une variable de tri & des tests statistiques).}
\usage{
tabph(tbx, nn = 2, nomv = "", normx = FALSE)
}
\arguments{
\item{tbx}{un tableau gtsummary}

\item{nn}{nombre de valeurs dans la variable de tri}

\item{nomv}{nom à afficher pour la variable de tri}

\item{normx}{si variables numériques normales TRUE}
}
\value{
un tableau fignolé
}
\description{
Réglages pour un tableau généré par gtsummary (avec une variable de tri & des tests statistiques).
}
\examples{
tabph(gtsummary::tbl_summary(patients[,1:4],by =sexe),  nn = 2, nomv = "Sexe", normx = FALSE)

}
