% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabreg.R
\name{tabregph}
\alias{tabregph}
\title{Tableau de régression}
\usage{
tabregph(reg, titre = "")
}
\arguments{
\item{titre}{Titre du tableau}

\item{ll}{Résultat une analyse en régression}
}
\value{
tableau
}
\description{
Tableau de régression
}
\examples{
ll <- glm(escarre~., data = patients, family = "binomial")
tabregph(reg = ll, titre = "Facteurs de risque d'escarre" )

}
