% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transang.R
\name{transangph}
\alias{transangph}
\title{TRANSFORMATION ANGULAIRE
Calcul des bornes sup et inf d'un intervalle de confiance par transformation angulaire.}
\usage{
transangph(vv, ld, pc = 95)
}
\arguments{
\item{vv}{nb evenement}

\item{ld}{taille echantillon}

\item{pc}{percent of IC (95 by default)}
}
\value{
binf = borne inferieur de l'IC,
bsup = borne superieure de l'IC,
nb = borne inf ; borne sup]
}
\description{
TRANSFORMATION ANGULAIRE
Calcul des bornes sup et inf d'un intervalle de confiance par transformation angulaire.
}
\examples{
transangph(vv = 55, ld = 100, pc = 95)

}
