% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioboxph.R
\name{vioboxph}
\alias{vioboxph}
\title{Graphique violon + boxplot}
\usage{
vioboxph(
  dfx,
  varx,
  varnum,
  titre = "",
  stit = "",
  titx = "",
  tity = "n",
  lab = ""
)
}
\arguments{
\item{dfx}{Tibble de référence}

\item{varx}{Variable catégorielle}

\item{varnum}{Variable numérique}

\item{titre}{Titre}

\item{stit}{Sous-titre}

\item{titx}{Titre de l'axe x}

\item{tity}{Titre de l'axe y}

\item{lab}{label}
}
\value{
graphique
}
\description{
Graphique violon + boxplot
}
\examples{
vioboxph(patients, admission, age, titre =" grands",
stit = "et grandes", titx = "Mode d'admission")

}
