% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barconfph.R
\name{barconfph}
\alias{barconfph}
\title{Barplot avec barres d'erreur
Trace un barplot avec les intervalles de confiance pour un tableau de deux variables numeriques}
\usage{
barconfph(varx, testx, valx = "oui", tit = "titre")
}
\arguments{
\item{varx}{cas a représenter}

\item{testx}{Variable de tri}

\item{valx}{valeur de testx a presenter}

\item{tit}{titre du graphique}
}
\value{
graphique
}
\description{
Barplot avec barres d'erreur
Trace un barplot avec les intervalles de confiance pour un tableau de deux variables numeriques
}
\examples{
xx <- c("A","A","B","A","B","C","A","C")
          yy <- c(rep("oui",4),rep("non",4))
         barconfph(varx=xx, testx= yy, valx = "non", tit = "essai")

}
