% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barconfph.R
\name{barconfph}
\alias{barconfph}
\title{Barplot avec barres d'erreur
Trace un barplot avec les intervalles de confiance pour une variable numerique et une variable factorielle de classification}
\usage{
barconfph(
  varx,
  testx,
  valx = "oui",
  titre = "",
  stitre = "",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{varx}{cas a représenter}

\item{testx}{Variable de tri}

\item{valx}{valeur de testx a presenter}

\item{titre}{Titre du graphique}

\item{stitre}{Soustitre du graphique}

\item{capt}{légende du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}

\item{lb}{label du graphique}
}
\value{
graphique
}
\description{
Barplot avec barres d'erreur
Trace un barplot avec les intervalles de confiance pour une variable numerique et une variable factorielle de classification
}
\examples{
xx <- factor(c("A","A","B","A","B","C","A","C"))
          yy <- factor(c(rep("oui",4),rep("non",4)))
         barconfph(varx=xx, testx= yy, valx = "non", tit = "essai", stitre = "", ang = 0)

}
