% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barconfph.R
\name{barconfph}
\alias{barconfph}
\title{Barconfph}
\usage{
barconfph(
  dfx,
  varnum,
  vartri,
  titre = "",
  stitre = "",
  tx = "",
  ty = "n",
  lab = "",
  cap = "",
  angle = 0
)
}
\arguments{
\item{dfx}{dataframe}

\item{varnum}{Variable numérique à afficher}

\item{vartri}{Variable de tri, factorielle}

\item{titre}{Titre du graphique (vide par défaut)}

\item{stitre}{Sous titre du graphique (vide par défaut)}

\item{tx}{Titre de l'axe des x (vide par défaut)}

\item{ty}{Titre de l'axe des x ("n" par défaut)}

\item{lab}{label  (vide par défaut) rarement utile}

\item{cap}{caption(texte écrit en bas du graphique)}

\item{angle}{d'affichage des labels de l'axe x s'ils ont trop longs}
}
\value{
un graphique
}
\description{
Tracé d'un graphique en barre représentant la moyenne avec son intervalle de confiance à
95 \% d'une valeur numérique découpé selon une variable factorielle.
}
\examples{
data("patients")
barconfph(dfx = patients, varnum = igs2, vartri = admission, 
titre = "IGS II vs provenance", stitre = "en ICU", tx = "Mode d'admission", ty = "IGS II", 
cap = "Texte écrit petit",angle = 0)

}
