% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barconfph.R
\name{barconfph}
\alias{barconfph}
\title{Barplot avec barres d'erreur d'une valeur
Trace un barplot avec les intervalles de confiance pour une valeur donnee
d'une variable factorielle
selon les classes d'une variable factorielle de classification}
\usage{
barconfph(
  dfx,
  varx,
  testx,
  valx = "oui",
  titre = "",
  stitre = "",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{dfx}{data.frame}

\item{varx}{variable a représenter (factoriel) y}

\item{testx}{Variable de tri (factoriel) x}

\item{valx}{valeur de testx a presenter}

\item{titre}{Titre du graphique}

\item{stitre}{Soustitre du graphique}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
graphique
}
\description{
Barplot avec barres d'erreur d'une valeur
Trace un barplot avec les intervalles de confiance pour une valeur donnee
d'une variable factorielle
selon les classes d'une variable factorielle de classification
}
\examples{
xx <- factor(c("A","A","B","A","B","C","A","C"))
          yy <- factor(c(rep("oui",4),rep("non",4)))
          dfe <- dplyr::tibble(xx,yy)
         barconfph(dfx = dfe,
                   varx=xx,
                   testx= yy,
                   valx = "non",
                   tit = "essai",
                   stitre = "",
                   ang = 0)

}
