% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bardeuxph.R
\name{bardeuxph}
\alias{bardeuxph}
\title{Barplot 2 variables factorielles
une variable de tri, une variable en \% par modalité de tri}
\usage{
bardeuxph(
  dfx,
  varp,
  vart,
  titre = "",
  stitre = "",
  ltitre = "",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{dfx}{data.frame}

\item{varp}{variable à exprimer en \% (factorielle)}

\item{vart}{variable factorielle de tri (axe des x)}

\item{titre}{Titre du graphique}

\item{stitre}{Sous titre}

\item{ltitre}{titre de la légende}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
un graphique
}
\description{
Barplot 2 variables factorielles
une variable de tri, une variable en \% par modalité de tri
}
\examples{
data("patients")
bardeuxph(patients,escarre, admission ,
titre = "Escarre & mode d'admission",
stitre = "\%", ltitre ="Escarre",
capt ="Escarre", lab = "aa", angle = 20)

}
