% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bardeuxph.R
\name{bardeuxph}
\alias{bardeuxph}
\title{Barplot 2 variables factorielles
une variable de tri, une variable en \% par modalité de tri}
\usage{
bardeuxph(varp, vart, titre = "")
}
\arguments{
\item{varp}{variable a exprimer en \% (factorielle)}

\item{vart}{variable factorielle de tri (axe des x)}

\item{titre}{Titre du graphique}
}
\value{
un graphique
}
\description{
Barplot 2 variables factorielles
une variable de tri, une variable en \% par modalité de tri
}
\examples{
aa <- c("a","a","b","c")
          bb <- c("1","2","3","3")
         bardeuxph(aa,bb,"essai")

}
