% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barouiph.R
\name{barouiph}
\alias{barouiph}
\title{Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, uniquement les "oui" selon les modalites de la variable de tri}
\usage{
barouiph(varp, vart, titre, oui = "oui")
}
\arguments{
\item{varp}{variable a traiter (factorielle)}

\item{vart}{variable de tri (afficher sur l'axe des x)}

\item{titre}{Titre du graphique}

\item{oui}{oui la valeur à afficher dans la variable varp}
}
\value{
un graphique
}
\description{
Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, uniquement les "oui" selon les modalites de la variable de tri
}
\examples{
aa <- c("a","a","b","c")
          bb <- c("oui","non","oui","oui")
         barouiph(bb,aa, titre = "essai", oui = "oui")


}
