% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barouiph.R
\name{barouiph}
\alias{barouiph}
\title{Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, en affichant uniquement les "oui" selon les modalitées de la variable de tri}
\usage{
barouiph(
  df,
  varx,
  vartri,
  oui = "oui",
  titre = "",
  stitre = "",
  ytitre = "\%",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{df}{data-frame}

\item{varx}{variable _ traiter (factorielle) le plus souvent binaire}

\item{vartri}{variable de tri (afficher sur l'axe des x)}

\item{oui}{oui la valeur _ afficher dans la variable varp}

\item{titre}{Titre du graphique}

\item{stitre}{Sous-titre du graphique}

\item{ytitre}{Titre de l'axe y (\%)}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
un graphique
}
\description{
Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, en affichant uniquement les "oui" selon les modalitées de la variable de tri
}
\examples{
barouiph(df = patients, varx = sexe, vartri = admission,
oui = "f", titre = "Mode d'admission",
stitre =" Pour les dames",
ytitre ="\%", capt ="Mode d'admission selon le sexe",
lab = "ab", angle = 0)

}
