% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barouiph.R
\name{barouiph}
\alias{barouiph}
\title{Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, uniquement les "oui" selon les modalitees de la variable de tri}
\usage{
barouiph(
  df,
  varx,
  vartri,
  oui = "oui",
  titre = "",
  stitre = "",
  ytitre = "\%",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{df}{data-frame}

\item{varx}{variable a traiter (factorielle) le plus souvent binaire}

\item{vartri}{variable de tri (afficher sur l'axe des x)}

\item{oui}{oui la valeur a afficher dans la variable varp}

\item{titre}{Titre du graphique}

\item{stitre}{Soustitre du graphique}

\item{ytitre}{Titre de l'axe y (\%)}

\item{capt}{legende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
un graphique
}
\description{
Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, uniquement les "oui" selon les modalitees de la variable de tri
}
\examples{
barouiph(df = esoph, varx = alcgp, vartri = tobgp,
oui = "0-39g/day", titre = "0-39g/day tt A",
stitre =" selon dose B",
ytitre ="\%", capt ="0-39g/day tt A selon dose B",
lab = "ab", angle = 0)

}
