% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barouiph.R
\name{barouiph}
\alias{barouiph}
\title{Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, uniquement les "oui" selon les modalités de la variable de tri}
\usage{
barouiph(
  varp,
  vart,
  oui = "oui",
  titre = "",
  stitre = "",
  capt = "",
  lab = "",
  angle = 0
)
}
\arguments{
\item{varp}{variable a traiter (factorielle) le plus souvent binaire}

\item{vart}{variable de tri (afficher sur l'axe des x)}

\item{oui}{oui la valeur à afficher dans la variable varp}

\item{titre}{Titre du graphique}

\item{stitre}{Soustitre du graphique}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
un graphique
}
\description{
Barplot uniquement les "oui" en \%
Trace un barplot d'une variable factorielle en oui/non, uniquement les "oui" selon les modalités de la variable de tri
}
\examples{
aa <- c("a","a","b","c","a")
          bb <- c("oui","non","oui","oui","oui")
         barouiph(aa,bb, titre = "essai", oui = "oui", angle = 0)


}
