% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barsimpleph.R
\name{barsimpleph}
\alias{barsimpleph}
\title{Barplot simple en \%
Trace un barplot d'une variable factorielle, axe des y en \%}
\usage{
barsimpleph(varx, titre = "", stitre = "", capt = "", lab = "", angle = 0)
}
\arguments{
\item{varx}{variable a traiter (factorielle)}

\item{titre}{Titre du graphique}

\item{stitre}{Soustitre du graphique}

\item{capt}{légende du graphique}

\item{lab}{label du graphique}

\item{angle}{angle affichage des valeurs de vart sur l'axe des x (0 par defaut)}
}
\value{
un graphique
}
\description{
Barplot simple en \%
Trace un barplot d'une variable factorielle, axe des y en \%
}
\examples{
barsimpleph(iris$Species, "Espece")
}
