% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bashaut.R
\name{bashaut}
\alias{bashaut}
\title{bashaut}
\usage{
bashaut(xx, bh)
}
\arguments{
\item{xx}{Une variable numérique}

\item{bh}{1 : borne basse, 2 : borne haute}
}
\value{
un nombre simple
}
\description{
Calcul de l'intervalle de confiance à 95 \% de la moyenne d'une variable numérique (utilisé par barconfph)
}
\examples{
data ("patients")
bashaut(xx =patients$age, bh = 1)


}
