% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaup.R
\name{beaup}
\alias{beaup}
\title{Petit p bien presente}
\usage{
beaup(varp, affp = 1)
}
\arguments{
\item{varp}{resultat du p}

\item{affp}{si 1 renvoie "p = 0,005" sinon "0,005"}
}
\value{
beaup , chaine de caractère
}
\description{
Petit p bien presente
}
\examples{
pp <- cor.test(iris$Sepal.Length,iris$Sepal.Width)
beaup(pp$p.value)

}
