% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiph.R
\name{bmiph}
\alias{bmiph}
\title{BMIPH
Coupe un vecteur BMI(numeric) (ou IMC) en facteurs avec les limites usuelles, en français ou en anglais}
\usage{
bmiph(bmi, lang = "fr")
}
\arguments{
\item{bmi}{BMI numérique}

\item{lang}{Langue de sortie (french or english (fr/eng)). default : french}
}
\value{
vecteur factoriel
}
\description{
BMIPH
Coupe un vecteur BMI(numeric) (ou IMC) en facteurs avec les limites usuelles, en français ou en anglais
}
\examples{
imc <- c(18,25,40,32)
bmiph(imc, "fr")
}
