% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debutph.R
\name{debutph}
\alias{debutph}
\title{Début du travail}
\usage{
debutph(fich)
}
\arguments{
\item{fich}{Fichier csv}
}
\value{
data.frame
}
\description{
Lit un csv, simplifie & normalise les titres des variables. Le csv a été créé par libreOffice avec les réglages par défaut.
Toutes les variables "caractère" sont transforées en facteur.
}
