% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debutph.R
\name{debutph}
\alias{debutph}
\title{Debut du travail}
\usage{
debutph(fich)
}
\arguments{
\item{fich}{Fichier csv}
}
\value{
data.frame
}
\description{
Lit un csv, simplifie & normalise les titres des variables. Le csv a ete cree par libreOffice avec les reglages by default.
Toutes les variables caractere sont transformees en facteur.
}
