% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debutph.R
\name{debutph}
\alias{debutph}
\title{Début du travail}
\usage{
debutph(fich, nax = c("", "na", "NA", " "))
}
\arguments{
\item{fich}{Fichier csv contenant les données}

\item{nax}{liste des valeurs à considérer comme NA}
}
\value{
data.frame
}
\description{
Lit un csv contenant les données, simplifie & normalise les titres des variables.
}
\details{
Le csv doit avoir les réglages par défaut (UTF-8, séparateur de champs : ",") soit les réglages par défaut de LibreOffice
.
Toutes les variables "caractère" sont transformées en facteur.
}
