% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexpk.R
\name{gexptabph}
\alias{gexptabph}
\title{Export de tableau}
\usage{
gexptabph(dfk, exp = FALSE, nomfich = "export", nomsheet = "x", lg = FALSE)
}
\arguments{
\item{dfk}{objet gtsummary}

\item{exp}{booleen.  TRUE : export vers ods}

\item{nomfich}{nom du classeur ods}

\item{nomsheet}{nom de la feuille dans le classeur ods}

\item{lg}{Booleeen. TRUE si sortie en longtable}
}
\value{
un tableau kbl prévu pour un export LaTeX
}
\description{
Sortie d'un tableau gtsummary en tableau kableExtra avec adaptation à la page et  comme feuille d'un tableur ods
}
\examples{
zz <- gtsummary::tbl_summary(iris)
          gexptabph(dfk = zz, exp = FALSE)

}
