% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmph.R, R/glmph2.R
\name{glmph}
\alias{glmph}
\title{Tableau de regresion logistique}
\usage{
glmph(dfx, vart, vars, bnom, titre = "Analyse multivariée", lab = "")

glmph(dfx, vart, vars, bnom, titre = "Analyse multivariée", lab = "")
}
\arguments{
\item{dfx}{nom du data-frame}

\item{vart}{nom de la variable dependante}

\item{vars}{vecteur des variables explicatives (garder l'ordre du data-frame)}

\item{bnom}{vecteur contenant les intitules corrects des variables (l'extraction des valeurs utiles est automatique)}

\item{titre}{titre du tableau}

\item{lab}{label}
}
\value{
tableau

tableau
}
\description{
Trace le tableau d'une regression logistique avec les intitules vrais des variables.

Trace le tableau d'une regression logistique avec les intitules vrais des variables
Affiche les tests bruts & ajustes
}
\examples{
irisz <- iris
          irisz$zz <- as.factor(rep(c("a","b"),75))
          glmph(dfx = "irisz",vart = "zz",vars = c("Sepal.Width","Species"),bnom = names(irisz))

irisz <- iris
          irisz$zz <- as.factor(rep(c("a","b"),75))
          glmph(dfx = "irisz",vart = "zz",vars = c("Sepal.Width","Species"),bnom = names(irisz))

}
