% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmph.R
\name{glmph}
\alias{glmph}
\title{Regression logistique, sortie en un tableau.}
\usage{
glmph(dfx, vart, vars, titre = "", lab = "")
}
\arguments{
\item{dfx}{Data-frame (chaine de caracteres)}

\item{vart}{Variable de tri (chaine de caracteres)}

\item{vars}{Variables a explorer (un vecteur)}

\item{titre}{Titre du tableau}

\item{lab}{Label du tableau}
}
\value{
tableau
}
\description{
Regression logistique, sortie en un tableau.
}
\examples{
counts <- c(18,17,15,20,10,20,25,13,12)
          outcome <- gl(3,1,9)
treatment <- as.factor(c(rep("o",4),rep("n",5)))
dff <- data.frame(treatment, outcome, counts)
glmph(dfx = "dff", vart =  "treatment", vars = c("counts", "outcome" ), titre = "test", lab = "tabtest")


}
