% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligfc.R
\name{ligfc}
\alias{ligfc}
\title{ligne facteurs tableau comparatif}
\usage{
ligfc(nom, varp, trip)
}
\arguments{
\item{nom}{nom de la ligne}

\item{varp}{variable explicative (facteur)}

\item{trip}{variable expliquee (facteur)}
}
\value{
ligne complète avec p
}
\description{
ligne facteurs tableau comparatif
}
\examples{
ligfc("petale",iris$Species,iris$Species)
}
