% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lignumc.R
\name{ligmedc}
\alias{ligmedc}
\title{ligne numerique  test non parametrique tableau comparatif}
\usage{
ligmedc(nom, varp, trip)
}
\arguments{
\item{nom}{nom de la ligne}

\item{varp}{variable explicative (numerique)}

\item{trip}{variable expliquee (facteur)}
}
\value{
ligne complète avec p
}
\description{
ligne numerique  test non parametrique tableau comparatif
}
\examples{
ligmedc("tt A", runif(88),c(rep("A",44),rep("B",44)))
}
