% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabrandom.R
\name{listrandph}
\alias{listrandph}
\title{Table de randomisation

Cr\00e9ation de tables de randomisation pour une étude sur plusieurs centres en blocs de taille variable. Sortie (fichiers .csv) sur une table globale & une table par centre.}
\usage{
listrandph(nbcent = 1, nbtrait = 2, nbcas)
}
\arguments{
\item{nbcent}{nombre de centres (default : 1)}

\item{nbtrait}{nombre de groupes (default : 2)}

\item{nbcas}{Nombre cas souhaites}
}
\value{
un csv global  + un csv par centre
}
\description{
Table de randomisation

Cr\00e9ation de tables de randomisation pour une étude sur plusieurs centres en blocs de taille variable. Sortie (fichiers .csv) sur une table globale & une table par centre.
}
\author{
Philippe MICHEL
}
