% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listrandph.R
\name{listrandph}
\alias{listrandph}
\title{Tables de randomisation multicentre avec taille des blocs aléatoire}
\usage{
listrandph(nbcent = 1, nbtrait = 2, nbcas)
}
\arguments{
\item{nbcent}{Nombre de centres}

\item{nbtrait}{Nombre de classes}

\item{nbcas}{Nombre total de cas souhaité}
}
\value{
1 fichier csv global & un fichier par centre
}
\description{
Tables de randomisation multicentre avec taille des blocs aléatoire
}
\examples{
listrandph(nbcent = 1, nbtrait = 2, nbcas = 100)

}
