% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyciph.R
\name{moyciph}
\alias{moyciph}
\title{Title Intervalle de Confiance par Bootstrap pour une moyenne
Calcul des  IC par bootstrap pour une variable numérique}
\usage{
moyciph(varx, ci = 95)
}
\arguments{
\item{varx}{la variable à étudier}

\item{ci}{Intervalle de confiance en \% (95 par défaut)}
}
\value{
bornes inf & sup de l'IC
}
\description{
Title Intervalle de Confiance par Bootstrap pour une moyenne
Calcul des  IC par bootstrap pour une variable numérique
}
\examples{
moyciph(iris$Sepal.Width, ci = 95)

}
