% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moys.R
\name{moys}
\alias{moys}
\title{Présentation moyenne  ecart-type}
\usage{
moys(x)
}
\arguments{
\item{x}{variable numerique}
}
\value{
moyenne +/- ecart-type
}
\description{
Présentation moyenne  ecart-type
}
\examples{
moys(iris$Sepal.Length)
}
