% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1ph.R
\name{tab1ph}
\alias{tab1ph}
\title{Tableau descriptif
Avec les intitules des variables en clair}
\usage{
tab1ph(
  dfx,
  nomv = "**",
  titre = "Tableau descriptif",
  label = "tabd",
  export = FALSE
)
}
\arguments{
\item{dfx}{Data.frame to explore}

\item{nomv}{liste des vrais noms des variables (names(dfx by default))}

\item{titre}{title of the table}

\item{label}{label of the table for RMarkdown or LaTeX}

\item{export}{si TRUE, export en CSV}
}
\description{
Tableau descriptif
Avec les intitules des variables en clair
}
\examples{
tab1ph(iris, nomv = names(iris), titre = "Table 1", label = "tabiiris")
}
