% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1ph.R
\name{tab1ph}
\alias{tab1ph}
\title{Tableau descriptif
Avec les intitules des variables en clair}
\usage{
tab1ph(
  dfx,
  nomv = NULL,
  nlignes = NULL,
  titre = "Tableau descriptif",
  label = "tabd",
  test = "moy",
  export = FALSE
)
}
\arguments{
\item{dfx}{Data.frame to explore}

\item{nomv}{liste des vrais noms des variables (names(dfx by default))}

\item{nlignes}{numéro des lignes a inclure dans le tableau}

\item{titre}{title of the table}

\item{label}{label of the table for RMarkdown or LaTeX}

\item{test}{présentation des résultats en moy/écart-type si ="moy", en médiane quartiles si = "med"}

\item{export}{si TRUE, export en CSV}
}
\description{
Tableau descriptif
Avec les intitules des variables en clair
}
\examples{
tab1ph(patients, nomv = names(patients), nlignes = c(2:8),
 titre = "Table 1", label = "tab1", test = "moy")

}
