% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab12ph.R
\name{tab1ph}
\alias{tab1ph}
\title{Tableau 1}
\usage{
tab1ph(dfx, titre = "", test = "med", capt = "")
}
\arguments{
\item{dfx}{Tibble}

\item{titre}{Nom à afficher au desu des colonnes de comparaison}

\item{test}{Valeurs numériques en moyenne ± écart-type (\code{moy}) ou médiane (quartiles) (\code{med})}

\item{capt}{Titre du graphique}
}
\value{
tableau
}
\description{
Génération d'un tableau simple pour une étude clinique avec tests unitaires pour les deux groupes. Aucun test.
}
\examples{
tabdph(dfx = patients, test = "med", titre = "Escarre", capt = "Tableau 1")

}
