% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1ph.R
\name{tab1ph}
\alias{tab1ph}
\title{Tableau descriptif
Averc les intitules des variables en clair}
\usage{
tab1ph(dfx, nomv, titre = "Tableau descriptif", label = "tabd")
}
\arguments{
\item{dfx}{Data.frame to explore}

\item{titre}{title of the table}

\item{label}{label of the table for RMarkdown or LaTeX}

\item{nom}{vecteur avec les vrais noms des variables}
}
\value{

}
\description{
Tableau descriptif
Averc les intitules des variables en clair
}
\examples{
tab1bph(iris,nom =names(iris),titre = "Table 1", label = "tabiiris")
}
