% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabc.R
\name{tabcph}
\alias{tabcph}
\title{Tableau comparatif (univarie)}
\usage{
tabcph(
  dfx,
  tri,
  nomv = "",
  titre = "Tableau comparatif",
  nomvar = "Traitement",
  lab = "tabcomp",
  fnote = "",
  test = "moy",
  export = FALSE
)
}
\arguments{
\item{dfx}{table a explorer}

\item{tri}{variable expliquee}

\item{nomv}{liste des vrais noms des variables}

\item{titre}{titre du tableau}

\item{nomvar}{nom de la variable de tri (nom des variables par defaut)}

\item{lab}{label}

\item{fnote}{note en bas de tableau}

\item{test}{parametrique "moy" ou non "med"}

\item{export}{si TRUE, export en CSV}
}
\description{
Tableau comparatif (univarie)
}
\examples{
library(gtsummary)
          tabcph(dfx = trial, tri = trt, nomv= names(trial), test = "moy")
}
