% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabc2ph.R
\name{tabcph}
\alias{tabcph}
\title{Tableau comparatif}
\usage{
tabcph(dfx, tri, test = "moy", titre = "", capt = "")
}
\arguments{
\item{dfx}{Tibble}

\item{tri}{Variable de tri, comparaison}

\item{test}{Test paramétrique & chi2 si = moy, sinon, test de Wilcoxon + fisher}

\item{titre}{Nom à afficher au dessus des colonnes de comparaison}

\item{capt}{Titre du graphique}
}
\value{
tableau
}
\description{
Génération d'un tableau comparatif pour une étude clinique avec tests unitaires pour les deux groupes. Choix du type de test (paramétriques ou non).
}
\examples{
tabcph(dfx = patients, tri = escarre, test = "moy", titre = "Escarre", capt = "Tableau 1")

}
