% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabc.R
\name{tabcph}
\alias{tabcph}
\title{Tableau comparatif (univarie)}
\usage{
tabcph(
  dfx,
  tri,
  nomv,
  titre = "Tableau comparatif",
  lab = "tabcomp",
  fnote = "",
  export = FALSE
)
}
\arguments{
\item{dfx}{table a esplorer}

\item{tri}{variable expliquee}

\item{nomv}{liste des vrais noms des variables}

\item{titre}{titre du tableau}

\item{lab}{label}

\item{fnote}{note en bas de tableau}

\item{export}{si TRUE, export en CSV}
}
\value{

}
\description{
Tableau comparatif (univarie)
}
\examples{
tabcph(iris, Species, levels(iris$Species), fnote = "demonstration")
}
