% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabc.R
\name{tabcph}
\alias{tabcph}
\title{Tableau comparatif (univarie)}
\usage{
tabcph(
  dfx,
  tri,
  nlignes = NULL,
  nomv = NULL,
  titre = "Tableau comparatif",
  nomvar = "Traitement",
  lab = "tabcomp",
  fnote = "",
  test = "moy",
  export = FALSE
)
}
\arguments{
\item{dfx}{table à explorer}

\item{tri}{variable expliquée}

\item{nlignes}{numéro des lignes à inclure dans le tableau}

\item{nomv}{liste des vrais noms des variables}

\item{titre}{titre du tableau}

\item{nomvar}{nom de la variable de tri (nom des variables par défaut)}

\item{lab}{label}

\item{fnote}{note en bas de tableau}

\item{test}{paramétrique : test = "moy"; non "paramétrique : tet = "med"}

\item{export}{si TRUE, export en CSV}
}
\description{
Tableau comparatif (univarie)
}
\examples{
tabcph(dfx = patients, tri = alite.7.j.av,
nomv= nn$noms,#####
titre = "Tableau comparatif", nlignes = c(2:8),
nomvar = "Alité avant",
lab = "tabcomp",
test = "moy",
fnote = "Résultats selon l'alitement avant", export = FALSE)

}
