% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transang.R
\name{transangph}
\alias{transangph}
\title{Calcul des bornes sup et inf d'un intervalle de confiance par transformation angulaire.}
\usage{
transangph(nb, total, pr = 95)
}
\arguments{
\item{nb}{nb evenement}

\item{total}{taille echantillon}

\item{pr}{percent of IC (95 by default)}
}
\value{
binf = borne inferieur de l'IC,
bsup = borne superieure de l'IC,
nb = borne inf ; borne sup]
}
\description{
Calcul des bornes sup et inf d'un intervalle de confiance par transformation angulaire.
}
\examples{
transangph(nb = 55, total = 100, pr = 95)

}
